# resources/swagger.yaml
openapi: 3.0.0
info:
  title: Автоматизированная система управления требованиями (АСУТр)
  description: |
    API для системы управления требованиями согласно ТЗ ГОСТ 34.602-2020.
    Система предоставляет функционал для управления требованиями, их версионированием,
    согласованием, трассировкой и интеграцией со смежными системами.
  version: 1.0.0
  contact:
    name: АО "МИФИСТ"
    url: https://mifist.example.com
    email: support@mifist.example.com
servers:
  - url: https://api.asut.example.com/v1
    description: Production server
  - url: http://localhost:8000/v1
    description: Development server
...
paths:
...
  /users/me/projects:
    get:
      tags:
        - Пользователи
      summary: Получить информацию о проектах пользователя (о себе)
      description: |
        Получение проектов текущего аутентифицированного пользователя.
        
        **Требует аутентификации:** Да
        **Права доступа:** Любой аутентифицированный пользователь
        **Возвращает:** Проекты пользователя и его роли на них
      operationId: get_user_projects
      responses:
        '200':
          description: Успешное получение проектов пользователя
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProjectsResponse'
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ================ Проекты ================
  /projects:
    get:
      tags: [Проекты]
      summary: Получить список проектов
      description: Получение списка проектов с возможностью фильтрации, в случае отсутствия - возвращать пустой массив. Проверять права на просмотр проекта, возыращать только доступные
      security:
        - bearerAuth: []
      parameters:
        - name: created_by
          in: query
          schema:
            type: integer
        - name: user_id
          required: true
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Список проектов
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectList'
        '403':
          description: Недостаточно прав для просмотр списка проектов
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags: [Проекты]
      summary: Создать проект
      description: Создание нового проекта, проверка прав на PROJECT_CREATE. В таблице project поле created_by_user_id заполняется по идентификатору текущего пользователя, role_id
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectCreate'
      responses:
        '201':
          description: Проект создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '403':
          description: Недостаточно прав для создания проекта
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Проект с таким названием уже существует
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /projects/{project_id}:
    get:
      tags: [Проекты]
      summary: Получить информацию о проекте
      security:
        - bearerAuth: []
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Информация о проекте
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '403':
          description: Недостаточно прав для чтения проекта
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      tags: [Проекты]
      summary: Обновить проект (возможно изменить описание проекта)
      security:
        - bearerAuth: []
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectUpdate'
      responses:
        '200':
          description: Проект обновлен
        '403':
          description: Недостаточно прав для обнлвления проекта
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags: [Проекты]
      summary: Удалить проект
      description: Полное удаление проекта со всеми связанными требованиями, обсуждениями и историей (в результате должна быть удалена информация из всех таблиц), для UI перед отправкой данного запроса обязано всплыть предупреждение с требованием подтвердить действие
      security:
        - bearerAuth: []
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Проект успешно удален
        '403':
          description: Недостаточно прав для удаления проекта
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Проект не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Проект нельзя удалить (есть активные зависимости)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /projects/{project_id}/requirements:
    get:
      tags: [Проекты, Требования]
      summary: Получить требования проекта
      description: Получение всех требований проекта
      security:
        - bearerAuth: []
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: integer
        - name: type
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Список требований проекта
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequirementListResponse'

  # ================ Требования ================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
        timestamp:
          type: string
          format: date-time
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        pages:
          type: integer

    # ================ Пользователи ================
...
    UserProjectRole:
      type: object
      required:
        - project_id
        - project_name
        - role_id
        - role_name
      properties:
        project_id:
          type: string
          format: uuid
          description: ID проекта
          example: "f9ec8db5-f1a6-4f16-aef2-a2b0e2da2f40"
        project_name:
          type: string
          description: Название проекта
          example: "Банковское мобильное приложение"
        role_id:
          type: string
          format: uuid
          description: ID роли
          example: "5bf2224d-9965-4928-8769-78fc19b014dc"
        role_name:
          type: string
          description: Название роли
          example: "PROJECT_MANAGER"
    UserProjectsResponse:
      type: object
      required:
        - project_roles
      properties:
        project_roles:
          type: array
          description: Роли пользователя на проектах
          items:
            $ref: '#/components/schemas/UserProjectRole'
...
    Project:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        created_at:
          type: string
          format: date-time
        status:
          type: string
          enum: [active, archived, draft]
        created_by_user_id:
          type: integer
    ProjectCreate:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          maxLength: 255
        description:
          type: string
        status:
          type: string
          enum: [active, draft]
          default: active
    ProjectUpdate:
      type: object
      properties:
        name:
          type: string
          maxLength: 255
        description:
          type: string
        status:
          type: string
          enum: [active, archived]
    ProjectList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Project'

    # ================ Требования ================
    Requirement:
      type: object
      properties:
        id:
          type: integer
        project_id:
          type: integer
        name:
          type: string
        type:
          type: string
        created_at:
          type: string
          format: date-time
        created_by_user_id:
          type: integer
        path:
          type: string
        depth:
          type: integer
        requirement_group_id:
          type: integer
    RequirementFull:
      allOf:
        - $ref: '#/components/schemas/Requirement'
        - type: object
          properties:
            content:
              $ref: '#/components/schemas/RequirementContent'
            workflow:
              $ref: '#/components/schemas/RequirementWorkflow'
            dependencies:
              type: array
              items:
                $ref: '#/components/schemas/Dependency'
...
    RequirementListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/RequirementFull'
        pagination:
          $ref: '#/components/schemas/Pagination'

    # ================ Содержимое требований ================
    RequirementContent:
      type: object
      properties:
        id:
          type: integer
        development_basis:
          type: string
        development_purpose:
          type: string
        description_text:
          type: string
        acceptance_criteria:
          type: string
        document_requires:
          type: string
        requirement_id:
          type: integer
...
    RequirementWorkflow:
      type: object
      properties:
        id:
          type: integer
        requirement_content_id:
          type: integer
        priority:
          type: integer
          minimum: 1
          maximum: 5
        status:
          type: string
          enum: [draft, review, approved, rejected, implemented, verified, archived]
...
    Dependency:
      type: object
      properties:
        id:
          type: integer
        target_requirement_id:
          type: integer
        source_requirement_id:
          type: integer
        type:
          type: string
          enum: [blocks, depends_on, related_to, duplicates, refines]
        strength:
          type: integer
          minimum: 1
          maximum: 5
        description:
          type: string
        source_requirement_group_id:
          type: integer
        target_requirement_group_id:
          type: integer
...